BUG: fix the library name. ENH: DCMTK module. ENH: DCMTK-IO module.
This is a copy of the work by bill ryan and drew wasem with a few minor bug fixes.
DCMTK is fetched from the CTK repository through the git submodule mechanism.
Modules/ThirParty/DCMTK contains the placeholder for the corresponding itk-module.
Modules/IO/DCMTK is the placeholder for the ITK IO Factory implementation and corresponding wrapping information.
 had to overide the pre-commit hook to add a submodule. Sorry.

Change-Id: Idcbd60d1adcb943d1ed252eba42e9f110bf60f1d